+++ crates/nwc/src/api/mod.rs
@@ -1,0 +1,15 @@
+mod get_balance;
+mod get_info;
+mod list_transactions;
+mod lookup_invoice;
+mod make_invoice;
+mod pay_invoice;
+mod pay_keysend;
+
+pub use self::get_balance::*;
+pub use self::get_info::*;
+pub use self::list_transactions::*;
+pub use self::lookup_invoice::*;
+pub use self::make_invoice::*;
+pub use self::pay_invoice::*;
+pub use self::pay_keysend::*;